refactor(renderer): move cross-group views to self/cross-group/; remove personal-views/
Step 5 of the renderer reorganization, final slice. The cross-group views
aggregate over all of an agent's groups — they are the Self's view across
its Groups, not view-hosting machinery — so cross-group-main and
cross-group-block move from layout/views/ to self/cross-group/.

personal-views/ — the last folder named for where it rendered rather than
what it was about — is now empty and removed.

Pure file moves + import-path updates; no behaviour change.
