+++ src/renderer/src/self/cross-group/cross-group-main.ts
@@ -7,7 +7,7 @@ import { IframeKind, RenderView } from '@theweave/api';
 import '@shoelace-style/shoelace/dist/components/spinner/spinner.js';
 import '@holochain-open-dev/elements/dist/elements/display-error.js';
 
-import './view-frame.js';
+import '../../layout/views/view-frame.js';
 import { MossStore } from '../../moss-store.js';
 import { mossStoreContext } from '../../context.js';
 import { mossStyles } from '../../shared-styles.js';
